requirements.txt
# Django Core
Django==5.0.6
python-dotenv==1.0.1

# Database
# Für MySQL/MariaDB (Standard für ISPConfig3)
PyMySQL==1.1.1

# Für SQLite (eingebaut in Python, keine extra Installation nötig)

# Optional: PostgreSQL (nur installieren wenn benötigt)
# psycopg2-binary==2.9.9  # Benötigt PostgreSQL auf Windows

dj-database-url==2.2.0

# REST API
djangorestframework==3.15.2
django-cors-headers==4.4.0

# Authentication & Authorization
msal==1.30.0

# Celery & Background Tasks
celery==5.4.0
django-celery-beat==2.6.0
django-celery-results==2.5.1
redis==5.0.8

# Email & Communication
anthropic==0.30.1

# Security
cryptography==42.0.8

# Monitoring & Error Tracking
sentry-sdk==2.8.0

# Utilities
Pillow==10.4.0
python-dateutil==2.9.0
pytz==2024.1

# Development & Testing
pytest==8.3.2
pytest-django==4.8.0
pytest-cov==5.0.0

# Production Server
gunicorn==22.0.0
whitenoise==6.7.0

# Code Quality (optional für Development)
# black==24.4.2
# flake8==7.1.0
# isort==5.13.2
